+++ scripts/dice/handlers/GenericRollHandler.js
@@ -1,4 +1,6 @@
 import { IRollHandler } from '../interfaces/IRollHandler.js';
+import { RollModePolicy } from '../RollModePolicy.js';
+import { Logger } from '../../utils/Logger.js';
 
 /**
  * Generic Roll Handler
@@ -11,7 +13,7 @@ export class GenericRollHandler extends IRollHandler {
     super();
     this.diceExtractor = diceExtractor;
     this.rollBuilder = rollBuilder;
-    this.logger = console;
+    this.logger = Logger;
   }
 
   /**
@@ -55,11 +57,14 @@ export class GenericRollHandler extends IRollHandler {
       if (buildFormula.isAdvantage) { 
         flavor += ' (Advantage)';
       }
-      else if (buildFormula.isDisadvantage) { 
+      else if (buildFormula.isDisadvantage) {
         flavor += ' (Disadvantage)';
       }
 
-      await roll.toMessage({ flavor, speaker });
+      if (typeof roll.toMessage === 'function') {
+        RollModePolicy.suppressAnimation();
+        await roll.toMessage({ flavor, speaker }, RollModePolicy.messageOptions(actor));
+      }
 
       this.logger.log(`DDB Sync | Processed ${rollType} roll for ${actor.name}`);
   }